+++ packages/common/src/core/schemas/internal/metrics/InitiativeUiSchemaMetrics.ts
@@ -195,6 +195,17 @@ export const buildUiSchema = async (
                   labelKey: "shared.fields.metrics.sourceLink.message.required",
                   allowShowBeforeInteract: true,
                 },
+                {
+                  type: "ERROR",
+                  keyword: "format",
+                  icon: true,
+                  labelKey: "shared.errors.urlFormat",
+                },
+                {
+                  type: "ERROR",
+                  keyword: "if",
+                  hidden: true,
+                },
               ],
             },
           },