+++ packages/common/src/events/api/orval/api/orval-events.ts
@@ -7,15 +7,48 @@
  */
 import { Awaited } from "../awaited-type";
 import { customClient } from "../custom-client";
+export interface IPagedRegistrationResponse {
+  items: IRegistration[];
+  nextStart: number;
+  total: number;
+}
+
+export enum RegistrationSort {
+  createdAt = "createdAt",
+  updatedAt = "updatedAt",
+  firstName = "firstName",
+  lastName = "lastName",
+  username = "username",
+}
 export type GetRegistrationsParams = {
   /**
    * Event id being registered for
    */
   eventId?: string;
   /**
-   * ArcGIS Online id for a user. Will always be extracted from the token unless service token is used.
+   * ArcGIS Online id for a user
    */
-  userAgoId?: string;
+  userId?: string;
+  /**
+   * comma separated string list of registration roles
+   */
+  role?: string;
+  /**
+   * comma separated string list of registration statuses
+   */
+  status?: string;
+  /**
+   * comma separated string list of registration types
+   */
+  type?: string;
+  /**
+   * latest ISO8601 updatedAt for the registrations
+   */
+  updatedAtBefore?: string;
+  /**
+   * earliest ISO8601 updatedAt for the registrations
+   */
+  updatedAtAfter?: string;
   /**
    * the max amount of registrations to return
    */
@@ -24,67 +57,15 @@ export type GetRegistrationsParams = {
    * the index to start at
    */
   start?: string;
-};
-
-export type GetEventsParams = {
   /**
-   * which relation fields to include in response
+   * property to sort results by
    */
-  include?: string;
+  sortBy?: RegistrationSort;
   /**
-   * latest ISO8601 start date-time for the events
-   */
-  startDateTimeBefore?: string;
-  /**
-   * earliest ISO8601 start date-time for the events
-   */
-  startDateTimeAfter?: string;
-  /**
-   * Comma separated string list of AttendanceTypes
-   */
-  attendanceTypes?: string;
-  /**
-   * Comma separated string list of categories
-   */
-  categories?: string;
-  /**
-   * comma separated string list of event statuses
-   */
-  status?: string;
-  /**
-   * Comma separated string list of tags
-   */
-  tags?: string;
-  /**
-   * string to match within an event title
-   */
-  title?: string;
-  /**
-   * the max amount of events to return
-   */
-  num?: string;
-  /**
-   * the index to start at
-   */
-  start?: string;
-  /**
-   * Event property to sort results by
-   */
-  sortBy?: EventSort;
-  /**
-   * sort results order desc or asc
+   * sort order desc or asc
    */
   sortOrder?: SortOrder;
 };
-
-export interface IUpdateRegistration {
-  /** Role of the user in the event */
-  role?: RegistrationRole;
-  /** Status of the registration */
-  status?: RegistrationStatus;
-  /** Attendance type for this registration */
-  type?: EventAttendanceType;
-}
 
 export interface ICreateRegistration {
   /** ArcGIS Online id for a user. Will always be extracted from the token unless service token is used. */
@@ -151,6 +132,12 @@ export interface IUpdateEvent {
   title?: string;
 }
 
+export interface IPagedEventResponse {
+  items: IEvent[];
+  nextStart: number;
+  total: number;
+}
+
 export enum SortOrder {
   asc = "asc",
   desc = "desc",
@@ -161,6 +148,57 @@ export enum EventSort {
   createdAt = "createdAt",
   updatedAt = "updatedAt",
 }
+export type GetEventsParams = {
+  /**
+   * Comma separated string list of relation fields to include in response
+   */
+  include?: string;
+  /**
+   * latest ISO8601 start date-time for the events
+   */
+  startDateTimeBefore?: string;
+  /**
+   * earliest ISO8601 start date-time for the events
+   */
+  startDateTimeAfter?: string;
+  /**
+   * Comma separated string list of AttendanceTypes
+   */
+  attendanceTypes?: string;
+  /**
+   * Comma separated string list of categories
+   */
+  categories?: string;
+  /**
+   * comma separated string list of event statuses
+   */
+  status?: string;
+  /**
+   * Comma separated string list of tags
+   */
+  tags?: string;
+  /**
+   * string to match within an event title
+   */
+  title?: string;
+  /**
+   * the max amount of events to return
+   */
+  num?: string;
+  /**
+   * the index to start at
+   */
+  start?: string;
+  /**
+   * Event property to sort results by
+   */
+  sortBy?: EventSort;
+  /**
+   * sort results order desc or asc
+   */
+  sortOrder?: SortOrder;
+};
+
 export interface IRegistrationPermission {
   canDelete: boolean;
   canEdit: boolean;
@@ -192,7 +230,7 @@ export interface IEvent {
   createdById: string | null;
   creator?: IUser;
   description: string | null;
-  editGroups: string[] | null;
+  editGroups: string[];
   endDateTime: string;
   geometry: IEventGeometry;
   id: string;
@@ -200,7 +238,7 @@ export interface IEvent {
   onlineMeetings?: IOnlineMeeting[];
   orgId: string;
   permission: IEventPermission;
-  readGroups: string[] | null;
+  readGroups: string[];
   recurrence: string | null;
   registrations?: IRegistration[];
   startDateTime: string;
@@ -223,6 +261,15 @@ export enum RegistrationRole {
   ORGANIZER = "ORGANIZER",
   ATTENDEE = "ATTENDEE",
 }
+export interface IUpdateRegistration {
+  /** Role of the user in the event */
+  role?: RegistrationRole;
+  /** Status of the registration */
+  status?: RegistrationStatus;
+  /** Attendance type for this registration */
+  type?: EventAttendanceType;
+}
+
 export enum EventStatus {
   PLANNED = "PLANNED",
   CANCELED = "CANCELED",
@@ -391,7 +438,7 @@ export const getEvents = (
   params?: GetEventsParams,
   options?: SecondParameter<typeof customClient>
 ) => {
-  return customClient<IEvent[]>(
+  return customClient<IPagedEventResponse>(
     { url: `/api/events/v1/events`, method: "GET", params },
     options
   );
@@ -452,7 +499,7 @@ export const getRegistrations = (
   params?: GetRegistrationsParams,
   options?: SecondParameter<typeof customClient>
 ) => {
-  return customClient<IRegistration[]>(
+  return customClient<IPagedRegistrationResponse>(
     { url: `/api/events/v1/registrations`, method: "GET", params },
     options
   );