+++ packages/common/src/items/unshare-item-from-groups.ts
@@ -9,8 +9,8 @@ import { IRequestOptions } from "@esri/arcgis-rest-request";
  * Unshare an item from a set of groups
  * @param {String} itemId Item Id to unshare from groups
  * @param {Array} groups Array of group id's from which the item will be unshared
- * @param {String} owner optional Owner username to determine which endpoint to hit
  * @param {IRequestOptions} requestOptions
+ * @param {String} owner optional Owner username to determine which endpoint to hit
  */
 export function unshareItemFromGroups(
   itemId: string,
@@ -33,9 +33,12 @@ export function unshareItemFromGroups(
       // individual call in another try/catch block to
       // make sure it catches before all promises finish
       try {
-        return await unshareItemWithGroup(opt);
+        const results = await unshareItemWithGroup(opt);
+        return results;
       } catch (err) {
-        throw Error(`Error unsharing item: ${itemId} with group: ${groupId}`);
+        throw new Error(
+          `Error unsharing item: ${itemId} with group: ${groupId}`
+        );
       }
     })
   );