+++ packages/common/test/events/_internal/EventUiSchemaCreate.test.ts
@@ -212,7 +212,7 @@ describe("EventUiSchemaCreate", () => {
                   type: "ERROR",
                   keyword: "format",
                   icon: true,
-                  labelKey: `myI18nScope.fields.onlineUrl.linkHrefFormatError`,
+                  labelKey: `shared.errors.urlFormat`,
                 },
               ],
             },